Refactor scheduler implementation.
In this commit I'm suggesting a few adjustments to the scheduler code that I
think will make it more readable, easier to maintain and easier to add new
features. I've tried as much as possible to stick to the original scheduler algorithm and the provided input/output schemas.

I've created a sepearate `output_formatter` module that handles the
formatting of the output from the `task_placer` for presentation to the
front-end. I think this will allow for more flexibility as we won't need to touch the core
logic of the scheduler when the requirements from the front-end change
(unless it's a major change). This is in keeping with
[Model-View-Controller architecture](https://en.wikipedia.org/wiki/Model%E2%80%93view%E2%80%93controller).

I've moved the logic of converting a goal into tasks to the goal struct itself.
Now the `task_generator`'s job has been greatly simplified; it's responsible for getting all the tasks
when given a number of goals. We still end up with a flat list of
tasks, not a hierarchy.

The algorithm is the same at it's core i.e. the concept of splitting the
available time into "slots" and giving tasks specific slots. However
I've tweaked the data structures involved a little bit, mainly to make it easier to
implement. Please check the following Google slides where I've tried to
explain how it works:

- [Simple Case](https://docs.google.com/presentation/d/1cfMOgFIVaYj-jKrUIuFADjmuomJhmSb-6SBYxoeptc8/edit?usp=sharing)
- [Multiple Days](https://docs.google.com/presentation/d/161PI1EnTg3yQf39PZBC2DX7alBdvzvesnghnJSx6SCM/edit?usp=sharing)
- [Long Duration](https://docs.google.com/presentation/d/1-9mjsbW2BsA1RjzJW4NelOEPMui3KWMV0MZBX3_l-Ys/edit?usp=sharing)

It should now be quite easy to adjust granularit as we desire. We can
generate slots of minutes as easily as we can slots of hours.

Currently, it's simply going from least flexible task to 2nd least flexible task
etc, but it's easy to adjust this in the `task_placer` to do least
flexible then the most flexible etc. Please note the tweaked algorithm does result in a
slightly different placement of the tasks for the basic test in issue #3 (see the above slides for a walk-through and the tests in `tests.rs`).

Although it's not yet handling things like
overlapping and failed placements of tasks, the `task_placer` has been greatly simplified.

One difference from before is that after being processed by the `task_placer`, each task holds a list of slots, plus other relevant information about the task. The `output_formatter` then handles the formatting of output data.

I've added some tests in `tests.rs` to test the new functions etc.

I haven't implemented a number of important things e.g.:
- Rewriting tests and adding more tests
- Error handling
- Handling failed placements of tasks
- Many other things

The changes in this commit are just suggestions! I've probably
overlooked some things and it may not be the way to go. I'll gladly
revert to the previous code if it is so desired.