Setup rust integration-tests infrastructure
This relates to issue #45.

To allow us to test the scheduler end-to-end using json files. The json
files will be located in the `js-tests` directory. Each separate test will
be in it's own directory with an `input.json` and `output.json` file.

For example the basic test is in `js-tests/basic-test/` and inside we
have an `input.json` and `output.json` that is used in for that specific
integration test.

The previous `tests.rs` has been renamed to `unit_tests.rs` to differentiate
it from the integration tests.