Make repetition-daily test pass.
Related to issue #43. The simple daily repetition is now passing.

Please note I've commented out the more complex repetition-daily-bounds in both the rust and deno
tests temporarily. The reason for this is that we may need to add one more set of parameters to the current agreed upon input format in order to be able to handle the scenario.