Create Error type.
To handle case where the output formatter is passed a task that doesn't
have a confirmed start/deadline.

Currently, this is never expected to happen,
as the task_placer only passes scheduled tasks with a confirmed start &
deadline to the output formatter. The calling function will handle this
by panicking with an appropriate message.

We can add more error variants to the Error type as we proceed.