Add some methods to Task object.
These methods remove some of the complexity from the `task_placer`, making
`task_placer` easier to read.

The next_start_deadline_combination is used to iterate through possible
start -> deadline combinations for a task. For example a task might have slots 10-15
and duration 2. So the first combination is 10-12, the next
combination is 11-13, then 12-14, etc.