Adjust output.json.
The output from the scheduler and original output.json differed in a few
places. I've adjusted the output.json to match the output of the
scheduler because the scheduler's output was the expected output as per
the algorithm.
For example on 2022-09-04, walk gets scheduled from 18hrs to 19hrs (not 17hrs
to 18hrs) and work gets scheduled from 14hrs to 18hrs (not 14hrs to 17hrs
then again from 19hrs to 20hrs).