Add shell script that downloads latest artifacts if present.
This script checks whether the ZinZen-Scheduler repo has a new version
of the wasm build (the contents of `pkg` directory), and if so, downloads the files
and replaces the local `pkg` directory with the newly downloaded files.

If the script is run on the front-end before tests are run, it will ensure the front-end has the latest binary of the scheduler in the `pkg` directory before the tests are run.

It should be run from the same location that the current `pkg` directory resides.

One side effect: it uses a file `artifact_id.txt` to store a unique id
for the last downloaded artifact.

Before running it, the environment needs a `ZINZENTOKEN` environmental variable holding a Github
access token that has read access to the repo:

```
export ZINZENTOKEN={Your github token}
```

Question: Would this be run locally while working on the front end or as part of the front end's CI?