Improve handling of optional, weekly, and impossible tasks.
Impossible: The time listed when reporting a conflict is the time of the
task that caused the conflict.

Optional: Optional tasks will not be split.

Weekly: Also, scheduler will now not attempt to schedule weekly tasks of the same goal on the same day.
See the test flex_repeat_1 to see an example of weekly tasks that would
otherwise combine on same day which is not what is intended when a user
specifies do something x times per week.